+++ native/src/embedded.rs
@@ -1,0 +1,1 @@
+pub static OMG_INTERPRETER_BC: &[u8] = include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), "/bytecode/interpreter.bc"));